backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
Pillow
sqlalchemy
pydantic
python-dotenv
face-recognition
dlib
